+++ lib/bloc/editor_bloc.dart
@@ -17,6 +17,7 @@ import 'dart:io';
 import 'dart:math';
 import 'dart:ui';
 
+import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:logger/logger.dart';
@@ -102,6 +103,29 @@ class FileIcons {
 }
 
 ///
+/// The Byte Order Mark, which was used during read / should be written for a file.
+///
+enum BomType {
+  /// File has no BOM
+  none([]),
+  /// File has a UTF8 BOM
+  utf8([0xEF, 0xBB, 0xBF]),
+  /// File has a UTF16 Big Endian Architecture BOM
+  utf16be([0xFE, 0xFF]),
+  /// File has a UTF32 Big Endian Architecture BOM
+  utf32be([0, 0, 0xFE, 0xFF]),
+  /// File has a UTF16 Little Endian Architecture BOM
+  utf16le([0xFF, 0xFE]),
+  /// File has a UTF32 Little Endian Architecture BOM
+  utf32le([0xFF, 0xFE, 0, 0]);
+  ///
+  /// The bytes representing the BomType
+  ///
+  final List<int> bytes;
+  const BomType(this.bytes);
+}
+
+///
 /// Represents one open file.
 ///
 class OpenFile {
@@ -131,6 +155,7 @@ class OpenFile {
   /// The character encoding.
   ///
   final Encoding encoding;
+  BomType bomType;
 
   ///
   /// The icon to represent this file.
@@ -202,6 +227,14 @@ class OpenFile {
     _changedListener!(this);
   }
 
+  ///
+  /// Run an arbitrary command changing the file.
+  ///
+  void runCommand(void Function() callback) {
+    callback();
+    _changedListener!(this);
+  }
+
   void onChanged(CodeLineEditingValue value) {
     if (controller.codeLines.equals(controller.preValue?.codeLines)) {
       final newCollapsed = value.selection.isCollapsed;
@@ -243,7 +276,11 @@ class OpenFile {
     required TextLineBreak lineBreak,
     this.readOnly = false,
     DateTime? modificationTime,
-    this.modified = false, this.encoding = utf8, required this.isNew, int? initialLineNumber}) {
+    this.modified = false,
+    this.encoding = utf8,
+    this.bomType = BomType.none,
+    required this.isNew,
+    int? initialLineNumber}) {
     language = Languages.singleton.modeForFilename(filename);
     _lastSavedText = text;
     this.modificationTime = modificationTime ?? DateTime.now();
@@ -376,7 +413,10 @@ class EditorBloc {
   Future<CommandResult> _saveFile(OpenFile fileHandle) async {
     try {
       final file = File(fileHandle.filename);
-      file.writeAsStringSync(fileHandle.controller.text, encoding: fileHandle.encoding);
+      final bytes = <int>[];
+      bytes.addAll(fileHandle.bomType.bytes);
+      bytes.addAll(fileHandle.encoding.encode(fileHandle.controller.text));
+      file.writeAsBytesSync(bytes);
       fileHandle.unchanged();
       return CommandResult(success: true, message: "Successfully saved ${fileHandle.filename}");
     } catch(ex) {
@@ -516,11 +556,19 @@ class EditorBloc {
     return CommandResult(success: true);
   }
 
-  ({Encoding encoding, TextLineBreak lineBreak}) _parseBytesToDetectEncoding(List<int> pData) {
+  ({Encoding encoding, TextLineBreak lineBreak, BomType bomType}) _parseBytesToDetectEncoding(List<int> pData) {
     var i = 0;
     final end = pData.length;
     int nLength;
     var lb = TextLineBreak.lf;
+    Encoding encoding = latin1;
+    var bomType = BomType.none;
+    for (var t in BomType.values) {
+      if (t.bytes.isNotEmpty && end > t.bytes.length && const ListEquality<int>().equals(t.bytes, pData.sublist(0, t.bytes.length))) {
+        bomType = t;
+        break;
+      }
+    }
     while (i < end) {
       int byte = pData[i++];
       if (byte == 13 && i < end && pData[i] == 10) {
@@ -541,46 +589,46 @@ class EditorBloc {
       }
       if (i + nLength >= end) {
         /* truncated string or invalid byte sequence */
-        return (encoding: latin1, lineBreak: lb);
+        return (encoding: encoding, lineBreak: lb, bomType: bomType);
       }
 
       /* Check continuation bytes: bit 7 should be set, bit 6 should be unset (b10xxxxxx). */
       for (i = 0; i < nLength; i++) {
         if ((pData[i] & 0xC0) != 0x80) {
-          return (encoding: latin1, lineBreak: lb);
+          return (encoding: encoding, lineBreak: lb, bomType: bomType);
         }
         if (nLength == 1) {
-          return (encoding: utf8, lineBreak: lb);
+          return (encoding: utf8, lineBreak: lb, bomType: bomType);
         } else if (nLength == 2) {
           /* 3 bytes sequence: U+0800..U+FFFF */
           int ch = ((pData[0] & 0x0f) << 12) + ((pData[1] & 0x3f) << 6) +
               (pData[2] & 0x3f);
           /* (0xff & 0x0f) << 12 | (0xff & 0x3f) << 6 | (0xff & 0x3f) = 0xffff, so ch <= 0xffff */
           if (ch < 0x0800) {
-            return (encoding: latin1, lineBreak: lb);
+            return (encoding: encoding, lineBreak: lb, bomType: bomType);
           }
           /* surrogates (U+D800-U+DFFF) are invalid in UTF-8: test if (0xD800 <= ch && ch <= 0xDFFF) */
           if ((ch >> 11) == 0x1b) {
-            return (encoding: latin1, lineBreak: lb);
+            return (encoding: encoding, lineBreak: lb, bomType: bomType);
           }
-          return (encoding: utf8, lineBreak: lb);
+          return (encoding: utf8, lineBreak: lb, bomType: bomType);
         } else if (nLength == 3) {
           /* 4 bytes sequence: U+10000..U+10FFFF */
           int ch = ((pData[0] & 0x07) << 18) + ((pData[1] & 0x3f) << 12) +
               ((pData[2] & 0x3f) << 6) + (pData[3] & 0x3f);
           if ((ch < 0x10000) || (0x10FFFF < ch)) {
-            return (encoding: latin1, lineBreak: lb);
+            return (encoding: encoding, lineBreak: lb, bomType: bomType);
           }
-          return (encoding: utf8, lineBreak: lb);
+          return (encoding: utf8, lineBreak: lb, bomType: bomType);
         }
         i += nLength;
       }
     }
-    return (encoding: latin1, lineBreak: lb);
+    return (encoding: encoding, lineBreak: lb, bomType: bomType);
   }
 
 
-  Future<({Encoding encoding, TextLineBreak lineBreak})> _detectEncoding(File file) async {
+  Future<({Encoding encoding, TextLineBreak lineBreak, BomType bomType})> _detectEncoding(File file) async {
     final size = min(4096, await file.length());
     final tester = file.openRead(0, size);
     return _parseBytesToDetectEncoding(await tester.first);
@@ -602,7 +650,13 @@ class EditorBloc {
       final stat = file.statSync();
       final readOnly = stat.readOnly;
       final result = await _detectEncoding(file);
-      String? text = file.readAsStringSync(encoding: result.encoding);
+      String? text;
+      if (result.bomType != BomType.none) {
+        var bytes = file.readAsBytesSync();
+        text = result.encoding.decode(bytes.sublist(result.bomType.bytes.length));
+      } else {
+        text = file.readAsStringSync(encoding: result.encoding);
+      }
       openFiles.remove(filename);
       openFiles.insert(0, filename);
       if (openFiles.length > 10) {