+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to the PKS Edit for Flutter will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [LATEST] - 2024-04-13
+## [LATEST] - 2024-04-25
 
 ### Added
 - Editing Settings will now save the settings right away for the next session. Generally editing settings was
@@ -16,4 +16,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Various new functions like Find..., Find&Replace... and Outdent and Indent were added. Also find again
   forward and backward will work in a way compatible to PKS EDIT Windows version.
 - L10N was improved and most texts are now available in English and German.
+- Byte Order Mark detection and handling was added.
+- New View Options (show/hide highlight, show/hide line numbers) were added.